pyproject.toml
[project]
name = "moonshine-flow"
version = "0.1.0"
description = "Push-to-talk transcription daemon for macOS using Moonshine"
readme = "README.md"
requires-python = ">=3.11,<3.12"
authors = [
  { name = "Moonshine Flow Contributors" }
]
dependencies = [
  "numpy>=2.0.0",
  "sounddevice>=0.5.1",
  "pynput>=1.7.7",
  "pyperclip>=1.9.0",
  "pydantic>=2.9.0",
  "tomli-w>=1.1.0",
  "moonshine-voice>=0.0.49",
  "pyobjc-core>=11.0; platform_system == 'Darwin'",
  "pyobjc-framework-Quartz>=11.0; platform_system == 'Darwin'",
  "pyobjc-framework-AVFoundation>=11.0; platform_system == 'Darwin'",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-mock>=3.14.0",
  "ruff>=0.8.0",
]

[project.scripts]
moonshine-flow = "moonshine_flow.cli:main"

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.uv]
default-groups = []

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["B008"]
